+++ core/utils/lume_config.ts
@@ -13,6 +13,7 @@ export const pluginNames = [
   "fff",
   "filter_pages",
   "google_fonts",
+  "gzip",
   "inline",
   "jsx",
   "jsx_preact",